+++ apps/podcast-api/src/main/java/nl/infosupport/agenttalks/podcast/service/PodcastGenerationException.java
@@ -1,0 +1,7 @@
+package nl.infosupport.agenttalks.podcast.service;
+
+public class PodcastGenerationException extends RuntimeException {
+    public PodcastGenerationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
