+++ apps/podcast-api/src/main/java/nl/infosupport/agenttalks/podcast/service/PodcastOrchestrator.java
@@ -8,12 +8,19 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.eclipse.microprofile.faulttolerance.Retry;
+import org.eclipse.microprofile.rest.client.inject.RestClient;
 import org.jboss.logging.Logger;
 
 import com.azure.storage.blob.BlobServiceClient;
 
+import io.quarkus.scheduler.Scheduled;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import nl.infosupport.agenttalks.podcast.clients.buzzsprout.BuzzsproutClient;
+import nl.infosupport.agenttalks.podcast.clients.buzzsprout.model.CreateEpisodeRequest;
+import nl.infosupport.agenttalks.podcast.clients.buzzsprout.model.CreateEpisodeResponse;
 import nl.infosupport.agenttalks.podcast.clients.content.ContentClient;
 import nl.infosupport.agenttalks.podcast.clients.content.model.ContentSubmission;
 import nl.infosupport.agenttalks.podcast.clients.content.model.CreatePodcastEpisode;
@@ -25,83 +32,111 @@ import nl.infosupport.agenttalks.podcast.model.PodcastScript;
 
 @ApplicationScoped
 public class PodcastOrchestrator {
-    
+
     @Inject
     ContentClient contentClient;
 
     @Inject
     BlobServiceClient blobServiceClient;
-    
+
     @Inject
     PodcastScriptGenerator podcastScriptGenerator;
-    
+
     @Inject
     AudioProcessor audioProcessor;
-    
+
     @Inject
-    PublishingService publishingService;
-    
+    @RestClient
+    BuzzsproutClient buzzsproutClient;
+
+    @ConfigProperty(name = "buzzsprout.podcast-id")
+    String podcastId;
+
     private static final Logger logger = Logger.getLogger(PodcastOrchestrator.class);
-    
+
+    @Scheduled(cron = "0 0 18 ? * FRI") // Every Friday at 18:00
+    public void runWeeklyPodcastGeneration() {
+        var currentDate = LocalDate.now();
+        var startDate = currentDate.minusDays(7); // One week ago
+
+        var processableSubmissions = contentClient.findProcessableSubmissions(startDate, currentDate);
+
+        logger.infof(
+                "Found %d processable submissions for the week starting %s",
+                processableSubmissions.size(), startDate);
+
+        if (!processableSubmissions.isEmpty()) {
+            logger.infof(
+                    "Triggering podcast generation for submissions from %s to %s",
+                    startDate, currentDate);
+
+            try {
+                generatePodcast(startDate, currentDate, processableSubmissions);
+                logger.info("Weekly podcast generation completed successfully");
+            } catch (Exception e) {
+                logger.errorf(e, "Weekly podcast generation failed for period %s to %s", startDate, currentDate);
+            }
+        } else {
+            logger.info("No processable submissions found for this week");
+        }
+    }
+
     public void generatePodcast(LocalDate startDate, LocalDate endDate, List<ContentSubmission> contentSubmissions) {
-        logger.infof("Starting podcast generation for %d submissions from %s to %s", 
-                    contentSubmissions.size(), startDate, endDate);
-        
+        logger.infof("Starting podcast generation for %d submissions from %s to %s",
+                contentSubmissions.size(), startDate, endDate);
+
         try {
             // Step 1: Lock content submissions
             lockContentSubmissions(contentSubmissions);
             logger.info("Content submissions locked successfully");
-            
+
             // Step 2: Generate podcast script
             PodcastScript script = generatePodcastScriptInternal(contentSubmissions);
             logger.infof("Generated podcast script with title: %s", script.title);
-            
+
             // Step 3: Generate audio for each script fragment
             var processableFragments = script.sections.stream()
                     .flatMap(section -> section.fragments.stream())
                     .collect(Collectors.toList());
-            
+
             var generatedFragments = processableFragments.stream()
                     .map(fragment -> audioProcessor.generateSpeech(fragment))
                     .collect(Collectors.toList());
             logger.infof("Generated %d audio fragments", generatedFragments.size());
-            
+
             // Step 4: Concatenate audio fragments
             var contentAudioFile = audioProcessor.concatenateAudioFragments(generatedFragments);
             logger.infof("Concatenated audio fragments into: %s", contentAudioFile);
-            
-            // Step 5: Mix final episode
-            var finalAudioFile = audioProcessor.mixPodcastEpisode(contentAudioFile);
-            logger.infof("Final podcast episode mixed: %s", finalAudioFile);
-            
-            // Step 6: Save episode metadata
-            PodcastEpisode episodeMetadata = savePodcastEpisode(script.title, finalAudioFile, contentSubmissions);
+
+            // Step 5: Save episode metadata
+            // Note: Intro/outro are handled by Buzzsprout during publishing
+            PodcastEpisode episodeMetadata = savePodcastEpisode(script.title, contentAudioFile, contentSubmissions);
             logger.infof("Saved episode metadata with episode number: %d", episodeMetadata.episodeNumber);
-            
-            // Step 7: Publish to Buzzsprout
-            publishingService.publishPodcastEpisode(
+
+            // Step 6: Publish to Buzzsprout
+            publishPodcastEpisode(
                     1, episodeMetadata.episodeNumber,
                     script.title,
                     episodeMetadata.description,
                     episodeMetadata.showNotes,
                     episodeMetadata.audioFilePath);
             logger.info("Published episode to Buzzsprout successfully");
-            
-            // Step 8: Mark submissions as processed
+
+            // Step 7: Mark submissions as processed
             markContentSubmissionsAsProcessed(contentSubmissions);
             logger.info("Marked content submissions as processed");
-            
+
             logger.info("Podcast generation completed successfully");
-            
+
         } catch (Exception e) {
             logger.errorf(e, "Podcast generation failed");
             throw new PodcastGenerationException("Failed to generate podcast", e);
         }
     }
-    
+
     private void lockContentSubmissions(List<ContentSubmission> contentSubmissions) {
         logger.infof("Locking %d content submissions", contentSubmissions.size());
-        
+
         for (var contentSubmission : contentSubmissions) {
             contentClient.markForProcessing(new MarkForProcessing(contentSubmission.id));
         }
@@ -109,7 +144,7 @@ public class PodcastOrchestrator {
 
     private void markContentSubmissionsAsProcessed(List<ContentSubmission> contentSubmissions) {
         logger.infof("Marking %d content submissions as processed", contentSubmissions.size());
-        
+
         for (var contentSubmission : contentSubmissions) {
             contentClient.markAsProcessed(new MarkAsProcessed(contentSubmission.id));
         }
@@ -117,7 +152,7 @@ public class PodcastOrchestrator {
 
     private PodcastScript generatePodcastScriptInternal(List<ContentSubmission> contentSubmissions) {
         logger.infof("Generating podcast script for %d content submissions", contentSubmissions.size());
-        
+
         var firstPodcastHost = PodcastHost.findByIndex(1);
         var secondPodcastHost = PodcastHost.findByIndex(2);
 
@@ -126,12 +161,12 @@ public class PodcastOrchestrator {
         }
 
         var combinedContent = combineContentSubmissions(contentSubmissions);
-        
+
         var script = podcastScriptGenerator.generatePodcastScript(
                 firstPodcastHost.name, secondPodcastHost.name,
                 firstPodcastHost.styleInstructions, secondPodcastHost.styleInstructions,
                 combinedContent);
-        
+
         logger.infof("Successfully generated script with title: %s", script.title);
         return script;
     }
@@ -148,10 +183,11 @@ public class PodcastOrchestrator {
         return combinedContent.toString();
     }
 
-    private PodcastEpisode savePodcastEpisode(String title, String audioFilePath,
+    @Retry(maxRetries = 3, delay = 2000, delayUnit = java.time.temporal.ChronoUnit.MILLIS)
+    PodcastEpisode savePodcastEpisode(String title, String audioFilePath,
             List<ContentSubmission> contentSubmissions) {
         logger.infof("Saving podcast episode: %s with audio file: %s", title, audioFilePath);
-        
+
         var audioFile = new File(audioFilePath);
         var episodesContainer = blobServiceClient.getBlobContainerClient("episodes");
 
@@ -221,7 +257,29 @@ public class PodcastOrchestrator {
 
         return description.toString();
     }
-    
+
+    @Retry(maxRetries = 3, delay = 1000, delayUnit = java.time.temporal.ChronoUnit.MILLIS)
+    String publishPodcastEpisode(int seasonNumber, int episodeNumber, String title, String description,
+            String showNotes, String audioFileUrl) {
+        logger.infof("Publishing podcast episode to Buzzsprout: %s (Season %d, Episode %d)",
+                title, seasonNumber, episodeNumber);
+
+        try {
+            CreateEpisodeRequest request = new CreateEpisodeRequest(
+                    seasonNumber, episodeNumber, title, description,
+                    showNotes, audioFileUrl);
+
+            CreateEpisodeResponse response = buzzsproutClient.createEpisode(podcastId, request);
+
+            logger.infof("Successfully published episode to Buzzsprout with ID: %d", response.id);
+
+            return response.id.toString();
+        } catch (Exception e) {
+            logger.errorf(e, "Failed to publish episode to Buzzsprout: %s", title);
+            throw new RuntimeException("Failed to publish episode to Buzzsprout", e);
+        }
+    }
+
     public static class PodcastGenerationException extends RuntimeException {
         public PodcastGenerationException(String message, Throwable cause) {
             super(message, cause);
