--- apps/podcast-api/src/test/java/nl/infosupport/agenttalks/podcast/service/PublishingServiceTest.java
@@ -1,75 +1,0 @@
-package nl.infosupport.agenttalks.podcast.service;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.when;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-import nl.infosupport.agenttalks.podcast.clients.buzzsprout.model.CreateEpisodeRequest;
-import nl.infosupport.agenttalks.podcast.clients.buzzsprout.model.CreateEpisodeResponse;
-import nl.infosupport.agenttalks.podcast.shared.PodcastTestFixtures;
-
-@QuarkusTest
-class PublishingServiceTest {
-    
-    private PublishingService service;
-    private PodcastTestFixtures.MockSetup mocks;
-
-    @BeforeEach
-    void setUp() {
-        mocks = PodcastTestFixtures.createMockSetup();
-        
-        service = new PublishingService();
-        service.buzzsproutClient = mocks.buzzsproutClient;
-        service.podcastId = "test-podcast-id";
-    }
-
-    @Test
-    void testPublishPodcastEpisodeSuccess() {
-        // Setup
-        String title = "Test Episode";
-        String description = "Test Description";
-        String showNotes = "Test Show Notes";
-        String audioFileUrl = "https://test.blob.core.windows.net/episodes/test.mp3";
-
-        CreateEpisodeResponse mockResponse = new CreateEpisodeResponse();
-        mockResponse.id = 12345L;
-        mockResponse.title = title;
-        mockResponse.publicUrl = "https://buzzsprout.com/test/12345";
-
-        when(mocks.buzzsproutClient.createEpisode(eq("test-podcast-id"), any(CreateEpisodeRequest.class)))
-                .thenReturn(mockResponse);
-
-        // Execute
-        String result = service.publishPodcastEpisode(
-                1, 1, title, description,
-                showNotes, audioFileUrl);
-
-        // Verify
-        assertEquals("12345", result);
-    }
-
-    @Test
-    void testPublishPodcastEpisodeFailure() {
-        // Setup
-        String title = "Test Episode";
-        String description = "Test Description";
-        String showNotes = "Test Show Notes";
-        String audioFileUrl = "https://test.blob.core.windows.net/episodes/test.mp3";
-
-        when(mocks.buzzsproutClient.createEpisode(eq("test-podcast-id"), any(CreateEpisodeRequest.class)))
-                .thenThrow(new RuntimeException("Buzzsprout API call failed"));
-
-        // Execute & Verify
-        RuntimeException exception = assertThrows(RuntimeException.class,
-                () -> service.publishPodcastEpisode(
-                        1, 1, title, description, showNotes, audioFileUrl));
-
-        assertEquals("Failed to publish episode to Buzzsprout", exception.getMessage());
-    }
-}
